Display Indeterminate Progress Bar on Item Load/Refresh
