+++ Files UWP/LoadingIndicator.cs
@@ -1,0 +1,32 @@
+﻿using System.ComponentModel;
+using Windows.UI.Xaml;
+
+namespace Files.Filesystem
+{
+    public class LoadingVisibility : INotifyPropertyChanged
+    {
+        public Visibility _isVisible;
+        public Visibility isVisible
+        {
+            get
+            {
+                return _isVisible;
+            }
+
+            set
+            {
+                if (value != _isVisible)
+                {
+                    _isVisible = value;
+                    NotifyPropertyChanged("isVisible");
+                }
+            }
+        }
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        private void NotifyPropertyChanged(string info)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(info));
+        }
+    }
+}
